fix: speed up install deps
