+++ package.json
@@ -10,7 +10,7 @@
   "scripts": {
     "build": "tsc --build && npm run copy-static && npm run build-bundles",
     "copy-static": "copyfiles 'src/browser/client-scripts/*' build",
-    "build-node-bundle": "esbuild ./src/bundle/index.ts --outdir=./build/src/bundle --bundle --format=cjs --platform=node --target=ES2021",
+    "build-node-bundle": "esbuild ./src/bundle/cjs/index.ts --outdir=./build/src/bundle/cjs --bundle --format=cjs --platform=node --target=ES2021",
     "build-browser-bundle": "node ./src/browser/client-scripts/build.js",
     "build-bundles": "concurrently -c 'auto' 'npm:build-browser-bundle' 'npm:build-node-bundle --minify'",
     "check-types": "tsc --project tsconfig.spec.json",
@@ -55,14 +55,9 @@
     "@gemini-testing/commander": "2.15.4",
     "@jspm/core": "2.0.1",
     "@puppeteer/browsers": "2.4.0",
-    "@types/debug": "4.1.12",
-    "@types/yallist": "4.0.4",
     "@vitest/spy": "2.1.4",
-    "@wdio/globals": "8.39.0",
     "@wdio/protocols": "8.38.0",
-    "@wdio/types": "8.39.0",
     "@wdio/utils": "8.39.0",
-    "@wdio/utils-cjs": "npm:@wdio/utils@7.26.0",
     "bluebird": "3.5.1",
     "chalk": "2.4.2",
     "clear-require": "1.0.1",
@@ -78,7 +73,6 @@
     "geckodriver": "4.5.0",
     "gemini-configparser": "1.4.1",
     "get-port": "5.1.1",
-    "glob-extra": "5.0.2",
     "import-meta-resolve": "4.0.0",
     "local-pkg": "0.4.3",
     "lodash": "4.17.21",
@@ -116,7 +110,7 @@
     "@commitlint/config-conventional": "^19.0.3",
     "@cspotcode/source-map-support": "0.8.0",
     "@sinonjs/fake-timers": "10.3.0",
-    "@swc/core": "1.3.40",
+    "@swc/core": "1.9.2",
     "@types/babel__code-frame": "7.0.6",
     "@types/babel__core": "7.20.5",
     "@types/bluebird": "3.5.38",
@@ -125,6 +119,7 @@
     "@types/chai-as-promised": "7.1.5",
     "@types/clear-require": "3.2.1",
     "@types/cli-progress": "3.11.6",
+    "@types/debug": "4.1.12",
     "@types/escape-string-regexp": "2.0.1",
     "@types/fs-extra": "11.0.4",
     "@types/lodash": "4.14.191",
@@ -137,8 +132,11 @@
     "@types/sinonjs__fake-timers": "8.1.2",
     "@types/urijs": "1.19.25",
     "@types/url-join": "4.0.3",
+    "@types/yallist": "4.0.4",
     "@typescript-eslint/eslint-plugin": "6.12.0",
     "@typescript-eslint/parser": "6.12.0",
+    "@wdio/types": "8.39.0",
+    "@wdio/utils-cjs": "npm:@wdio/utils@7.26.0",
     "aliasify": "1.9.0",
     "app-module-path": "2.2.0",
     "browserify": "13.3.0",
@@ -151,6 +149,7 @@
     "eslint": "8.25.0",
     "eslint-config-gemini-testing": "2.8.0",
     "eslint-config-prettier": "8.7.0",
+    "glob-extra": "5.0.2",
     "husky": "0.11.4",
     "jsdom": "^24.0.0",
     "jsdom-global": "3.0.2",
