+++ src/browser/existing-browser.js
@@ -5,7 +5,7 @@ const url = require("url");
 const Promise = require("bluebird");
 const _ = require("lodash");
 const webdriverio = require("webdriverio");
-const { sessionEnvironmentDetector } = require("@wdio/utils-cjs");
+const { sessionEnvironmentDetector } = require("../bundle/@wdio-utils");
 const { Browser } = require("./browser");
 const commandsList = require("./commands");
 const Camera = require("./camera");
