+++ src/bundle/@wdio-utils.ts
@@ -1,0 +1,5 @@
+/* eslint-disable @typescript-eslint/no-var-requires */
+const bundle = require("./cjs");
+
+export const sessionEnvironmentDetector: typeof import("@wdio/utils-cjs").sessionEnvironmentDetector =
+    bundle.wdioUtils.sessionEnvironmentDetector;
