+++ src/bundle/cjs/index.ts
@@ -1,0 +1,3 @@
+export * from "./test-transformer";
+export * from "./vendors/@wdio/utils-cjs";
+export * from "./vendors/glob-extra";
