+++ src/bundle/cjs/vendors/@wdio/utils-cjs.ts
@@ -1,0 +1,2 @@
+import * as wdioUtilsFns from "@wdio/utils-cjs";
+export const wdioUtils = wdioUtilsFns;
