+++ src/bundle/cjs/vendors/glob-extra.ts
@@ -1,0 +1,2 @@
+import * as globExtraFns from "glob-extra";
+export const globExtra = globExtraFns;
