--- src/bundle/constants.ts
@@ -1,2 +1,0 @@
-export const TRANSFORM_EXTENSIONS = [".jsx", ".cjsx", ".mjsx", ".tsx", ".ctsx", ".mtsx"];
-export const JS_EXTENSION_RE = /^\.([cm]?[tj]sx?|json)$/;
