--- src/bundle/index.ts
@@ -1,2 +1,0 @@
-export * from "./test-transformer";
-export * from "./constants";
